# backend/api-docs.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Japan SSW Platform API",
    "version": "1.0.0",
    "description": "REST API for Japan Specified Skilled Worker (SSW) job platform - connects Filipino jobseekers with Japanese employers",
    "contact": {
      "name": "API Support",
      "email": "support@japanssw.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development server"
    },
    {
      "url": "https://api.japanssw.com",
      "description": "Production server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT token obtained from login"
      }
    },
    "schemas": {
...
      "Company": {
        "type": "object",
        "properties": {
          "_id": {
            "type": "string"
          },
          "name": {
            "type": "string",
            "example": "Tech Innovation Corp"
          },
          "slug": {
            "type": "string",
            "example": "tech-innovation-corp"
          },
          "industry": {
            "type": "string",
            "example": "Manufacturing"
          },
          "size": {
            "type": "string",
            "example": "201-500"
          },
          "description": {
            "type": "string"
          },
          "location": {
            "type": "object",
            "properties": {
              "prefecture": {
                "type": "string"
              },
              "city": {
                "type": "string"
              },
              "postalCode": {
                "type": "string"
              }
            }
          },
          "isVerified": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
...
  "paths": {
...
    "/api/v1/companies": {
      "get": {
        "summary": "Get all companies",
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "schema": {
              "type": "integer",
              "default": 1
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 10
            }
          },
          {
            "in": "query",
            "name": "industry",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Companies retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "companies": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Company"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new company",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "name",
                  "industry"
                ],
                "properties": {
                  "name": {
                    "type": "string",
                    "example": "Tech Innovation Corp"
                  },
                  "industry": {
                    "type": "string",
                    "example": "Manufacturing"
                  },
                  "size": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  },
                  "location": {
                    "type": "object",
                    "properties": {
                      "prefecture": {
                        "type": "string"
                      },
                      "city": {
                        "type": "string"
                      }
                    }
                  },
                  "website": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Company created"
          }
        }
      }
    },
    "/api/v1/companies/{idOrSlug}": {
      "get": {
        "summary": "Get company by ID or slug",
        "tags": [
          "Companies"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "idOrSlug",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID or slug"
          }
        ],
        "responses": {
          "200": {
            "description": "Company retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Company"
                }
              }
            }
          },
          "404": {
            "description": "Company not found"
          }
        }
      }
    },
    "/api/v1/companies/{id}": {
      "put": {
        "summary": "Update company",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company updated"
          }
        }
      },
      "delete": {
        "summary": "Delete company",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company deleted"
          }
        }
      }
    },
    "/api/v1/companies/{id}/admins": {
      "post": {
        "summary": "Add company admin",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin added"
          }
        }
      }
    },
    "/api/v1/companies/{id}/admins/{userId}": {
      "delete": {
        "summary": "Remove company admin",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Admin removed"
          }
        }
      }
    },
    "/api/v1/companies/{id}/verify": {
      "put": {
        "summary": "Verify company (Admin only)",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Company verified"
          }
        }
      }
    },
    "/api/v1/companies/admin/stats": {
      "get": {
        "summary": "Get company statistics (Admin only)",
        "tags": [
          "Companies"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Company statistics"
          }
        }
      }
    },
...
    "/api/v1/jobs/company/{companyId}": {
      "get": {
        "summary": "Get all jobs for a specific company",
        "tags": [
          "Jobs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "companyId",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Company ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Company jobs retrieved"
          }
        }
      }
    },
...
  }
}